// src/app/E-Training_API/models/auth-response-view-model.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AuthResponseViewModel {
  Email?: string | null;
  ExpiresAt?: string;
  Nom?: string | null;
  Prenom?: string | null;
  Role?: string | null;
  Token?: string | null;
  UserId?: string | null;
}